Move fetch logic into repository
